added EnvironmentNetworkBase
