+++ environment-api/src/main/java/com/sequenceiq/environment/api/v1/environment/model/request/EnvironmentNetworkRequest.java
@@ -1,64 +1,14 @@
 package com.sequenceiq.environment.api.v1.environment.model.request;
 
-import java.util.Set;
-
-import javax.validation.constraints.Size;
-
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
-import com.sequenceiq.environment.api.doc.environment.EnvironmentModelDescription;
-import com.sequenceiq.environment.api.v1.environment.model.EnvironmentNetworkAwsParams;
-import com.sequenceiq.environment.api.v1.environment.model.EnvironmentNetworkAzureParams;
+import com.sequenceiq.environment.api.v1.environment.model.base.EnvironmentNetworkBase;
 
 import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
 
 @ApiModel(value = "EnvironmentNetworkV1Request")
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(Include.NON_NULL)
-public class EnvironmentNetworkRequest {
-    @ApiModelProperty(value = EnvironmentModelDescription.SUBNET_IDS, required = true)
-    private Set<String> subnetIds;
-
-    @Size(max = 255)
-    private String networkCidr;
-
-    @ApiModelProperty(EnvironmentModelDescription.AWS_SPECIFIC_PARAMETERS)
-    private EnvironmentNetworkAwsParams aws;
-
-    @ApiModelProperty(EnvironmentModelDescription.AZURE_SPECIFIC_PARAMETERS)
-    private EnvironmentNetworkAzureParams azure;
-
-    public Set<String> getSubnetIds() {
-        return subnetIds;
-    }
-
-    public void setSubnetIds(Set<String> subnetIds) {
-        this.subnetIds = subnetIds;
-    }
-
-    public String getNetworkCidr() {
-        return networkCidr;
-    }
-
-    public void setNetworkCidr(String networkCidr) {
-        this.networkCidr = networkCidr;
-    }
-
-    public EnvironmentNetworkAwsParams getAws() {
-        return aws;
-    }
-
-    public void setAws(EnvironmentNetworkAwsParams aws) {
-        this.aws = aws;
-    }
-
-    public EnvironmentNetworkAzureParams getAzure() {
-        return azure;
-    }
-
-    public void setAzure(EnvironmentNetworkAzureParams azure) {
-        this.azure = azure;
-    }
+public class EnvironmentNetworkRequest extends EnvironmentNetworkBase {
 }
